refactor(proxy): modularize app and harden runtime paths
