+++ .env.example
@@ -16,6 +16,30 @@
 # 'Authorization' header as a Bearer token (e.g., "Authorization: Bearer YOUR_PROXY_API_KEY").
 #PROXY_API_KEY="YOUR_PROXY_API_KEY"
 
+# ------------------------------------------------------------------------------
+# | [SECURITY] CORS Configuration                                              |
+# ------------------------------------------------------------------------------
+#
+# Control Cross-Origin Resource Sharing (CORS) for browser-based clients.
+# For production, set explicit origins instead of wildcard (*).
+#
+#PROXY_CORS_ORIGINS="https://yourdomain.com,https://app.yourdomain.com"
+#PROXY_CORS_CREDENTIALS="false"
+#
+# Default: PROXY_CORS_ORIGINS="*" (all origins allowed - INSECURE for production)
+# Default: PROXY_CORS_CREDENTIALS="false"
+
+# ------------------------------------------------------------------------------
+# | [PERFORMANCE] Model List Cache                                             |
+# ------------------------------------------------------------------------------
+#
+# TTL for the /v1/models endpoint cache. Models are cached per-provider to
+# avoid repeated API calls. Cache is invalidated on credential refresh.
+#
+#MODEL_LIST_CACHE_TTL="300"
+#
+# Default: 300 seconds (5 minutes)
+
 
 # ------------------------------------------------------------------------------
 # | [API KEYS] Provider API Keys                                               |
