+++ src/proxy_app/models.py
@@ -1,0 +1,111 @@
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2026 Mirrowel
+
+"""
+Pydantic models for the proxy application.
+
+This module contains all request/response models used by the API endpoints.
+"""
+
+import time
+from typing import List, Optional, Union
+from pydantic import BaseModel, ConfigDict, Field
+
+
+class EmbeddingRequest(BaseModel):
+    """Request model for embedding endpoint."""
+    model: str
+    input: Union[str, List[str]]
+    input_type: Optional[str] = None
+    dimensions: Optional[int] = None
+    user: Optional[str] = None
+
+
+class ModelCard(BaseModel):
+    """Basic model card for minimal response."""
+    id: str
+    object: str = "model"
+    created: int = Field(default_factory=lambda: int(time.time()))
+    owned_by: str = "Mirro-Proxy"
+
+
+class ModelCapabilities(BaseModel):
+    """Model capability flags."""
+    tool_choice: bool = False
+    function_calling: bool = False
+    reasoning: bool = False
+    vision: bool = False
+    system_messages: bool = True
+    prompt_caching: bool = False
+    assistant_prefill: bool = False
+
+
+class EnrichedModelCard(BaseModel):
+    """Extended model card with pricing and capabilities."""
+    id: str
+    object: str = "model"
+    created: int = Field(default_factory=lambda: int(time.time()))
+    owned_by: str = "unknown"
+    # Pricing (optional - may not be available for all models)
+    input_cost_per_token: Optional[float] = None
+    output_cost_per_token: Optional[float] = None
+    cache_read_input_token_cost: Optional[float] = None
+    cache_creation_input_token_cost: Optional[float] = None
+    # Limits (optional)
+    max_input_tokens: Optional[int] = None
+    max_output_tokens: Optional[int] = None
+    context_window: Optional[int] = None
+    # Capabilities
+    mode: str = "chat"
+    supported_modalities: List[str] = Field(default_factory=lambda: ["text"])
+    supported_output_modalities: List[str] = Field(default_factory=lambda: ["text"])
+    capabilities: Optional[ModelCapabilities] = None
+    # Debug info (optional)
+    _sources: Optional[List[str]] = None
+    _match_type: Optional[str] = None
+
+    model_config = ConfigDict(extra="allow")  # Allow extra fields from the service
+
+
+class ModelList(BaseModel):
+    """List of models response."""
+    object: str = "list"
+    data: List[ModelCard]
+
+
+class EnrichedModelList(BaseModel):
+    """List of enriched models with pricing and capabilities."""
+    object: str = "list"
+    data: List[EnrichedModelCard]
+
+
+class CostEstimateRequest(BaseModel):
+    """Request model for cost estimation endpoint."""
+    model: str
+    prompt_tokens: int = 0
+    completion_tokens: int = 0
+    cache_read_tokens: int = 0
+    cache_creation_tokens: int = 0
+
+
+class CostEstimateResponse(BaseModel):
+    """Response model for cost estimation endpoint."""
+    model: str
+    cost: Optional[float] = None
+    currency: str = "USD"
+    pricing: dict = Field(default_factory=dict)
+    source: Optional[str] = None
+
+
+class TokenCountRequest(BaseModel):
+    """Request model for token count endpoint."""
+    model: str
+    messages: List[dict]
+
+
+class RefreshQuotaStatsRequest(BaseModel):
+    """Request model for quota stats refresh endpoint."""
+    action: str = "reload"  # "reload" or "force_refresh"
+    scope: str = "all"  # "all", "provider", or "credential"
+    provider: Optional[str] = None
+    credential: Optional[str] = None
