+++ src/proxy_app/routes/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2026 Mirrowel
+
+"""Route modules for the proxy application."""
